This is for my younger siblings Chris, Christine, Melissa and Catherine.
B/c being afraid of things you can't *easily* explain is your "problem".

Trying to figure out how to understand that "problem" is the "solution". 

Some, if not most of the time you'll find others are also trying to put an *honest* effort into finding a "solution" too! :) 

Love, 
Bro